+++ kotti_grid/static/js/script.js
@@ -31,19 +31,23 @@ $(function() {
             field = window.opener.document.getElementById(field)
             var url = field.getAttribute('data-url');
             var use = field.getAttribute('data-use');
+            var custom_text = field.getAttribute('data-custom-text');
             var extra_style = field.getAttribute('data-extra-style');
             if (url !== null) $("input[name=url]").val(url);
             if (use !== null) $("select[name=use]").val(use);
             if (extra_style !== null) $("input[name=extra_style]").val(extra_style);
+            if (custom_text !== null) tinyMCE.activeEditor.setContent(custom_text);
         },
         submit: function() {
             var url = $("input[name=url]").val();
             var use = $("select[name=use]").val();
+            var custom_text = tinyMCE.activeEditor.getContent();
             var extra_style = $("input[name=extra_style]").val();
             var field = window.opener.document.getElementById('grid-fieldname').value;
             field = window.opener.document.getElementById(field)
             field.setAttribute('data-url', url);
             field.setAttribute('data-use', use);
+            field.setAttribute('data-custom-text', custom_text);
             field.setAttribute('data-extra-style', extra_style);
             window.opener.reload_tile(field);
             return window.close();
@@ -55,10 +59,12 @@ $(function() {
         var data = {};
         var url = tile.attr('data-url');
         var use = tile.attr('data-use');
+        var custom_text = tile.attr('data-custom-text');
         var extra_style = tile.attr('data-extra-style');
         var size_x = tile.attr('data-sizex');
         data['url'] = url;
         data['use'] = use;
+        data['custom_text'] = custom_text
         data['extra_style'] = extra_style
         var uri = '/@@tile-content';
         if(size_x !== undefined) {