+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 * Fix the responsive behavior of the grid tiles if the browser window is resized.
 * Setting the permission for the settings view to view instead of edit.
 * Raise test coverage.
+* Added a rich text field custom text to add free tile content.
 
 
 0.1a2(2013-05-03)